+++ util.h
@@ -3,7 +3,7 @@
 
 #include <cstdint>
 #include <vector>
-uint64_t intToBitboard(int num);
-int bitBoardToInt(uint64_t bitboard);  // Ensure proper casing
-std::vector<int> bitboardToVector(uint64_t bitboard);
+uint64_t intToBitboard(uint8_t num);
+uint8_t bitBoardToInt(uint64_t bitboard);  // Ensure proper casing
+std::vector<uint8_t> bitboardToVector(uint64_t bitboard);
 #endif // UTIL_H
